external_frontend/requirements.txt
django-app-settings==0.0.1
django-external-frontend==0.0.1
watchdog==0.7.1
PyYaml==3.11
django-pipeline==1.3.24
django-pipeline-compass==0.1.5
git+git:github.com/ludwigkraatz/django-introspective-api.git@f22120b83062cf544ff9383104da875f4da4508b